Refactor: JWT 리팩토링
JWT 패키지 리팩토링
UserRepository 직접 사용 X -> 인프라 계층에서 구현
ErrorCode 최상단으로 분리
