+++ src/main/java/com/comeon/backend/jwt/application/UserSimple.java
@@ -1,0 +1,13 @@
+package com.comeon.backend.jwt.application;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserSimple {
+
+    private Long userId;
+    private String nickname;
+    private String authorities;
+}
