+++ src/main/java/com/comeon/backend/jwt/application/UserSimpleService.java
@@ -1,0 +1,6 @@
+package com.comeon.backend.jwt.application;
+
+public interface UserSimpleService {
+
+    UserSimple loadUserSimple(Long userId);
+}
